+++ type/noncommutative/int64_codec.go
@@ -26,7 +26,7 @@ import (
 	"github.com/ethereum/go-ethereum/rlp"
 )
 
-func (this *Int64) Size() uint32 {
+func (this *Int64) Size() uint64 {
 	return 8 // 8 bytes
 }
 func (this *Int64) Encode() []byte {
